requirements.txt
numpy
pandas
tqdm
datasets
pythainlp
sentence-transformers
faiss-gpu
requests
python-crfsuite
